+++ cpp/pnp/quaternion.h
@@ -47,8 +47,8 @@ inline Eigen::Vector4f rotmat_to_quat(const RowMajorMatrix3f &R) {
 
 inline Eigen::Vector4f quat_multiply(const Eigen::Vector4f &qa,
                                      const Eigen::Vector4f &qb) {
-    const double qa1 = qa(0), qa2 = qa(1), qa3 = qa(2), qa4 = qa(3);
-    const double qb1 = qb(0), qb2 = qb(1), qb3 = qb(2), qb4 = qb(3);
+    const Float qa1 = qa(0), qa2 = qa(1), qa3 = qa(2), qa4 = qa(3);
+    const Float qb1 = qb(0), qb2 = qb(1), qb3 = qb(2), qb4 = qb(3);
 
     return Eigen::Vector4f(qa1 * qb1 - qa2 * qb2 - qa3 * qb3 - qa4 * qb4,
                            qa1 * qb2 + qa2 * qb1 + qa3 * qb4 - qa4 * qb3,
@@ -58,12 +58,12 @@ inline Eigen::Vector4f quat_multiply(const Eigen::Vector4f &qa,
 
 inline Eigen::Vector3f quat_rotate(const Eigen::Vector4f &q,
                                    const Eigen::Vector3f &p) {
-    const double q1 = q(0), q2 = q(1), q3 = q(2), q4 = q(3);
-    const double p1 = p(0), p2 = p(1), p3 = p(2);
-    const double px1 = -p1 * q2 - p2 * q3 - p3 * q4;
-    const double px2 = p1 * q1 - p2 * q4 + p3 * q3;
-    const double px3 = p2 * q1 + p1 * q4 - p3 * q2;
-    const double px4 = p2 * q2 - p1 * q3 + p3 * q1;
+    const Float q1 = q(0), q2 = q(1), q3 = q(2), q4 = q(3);
+    const Float p1 = p(0), p2 = p(1), p3 = p(2);
+    const Float px1 = -p1 * q2 - p2 * q3 - p3 * q4;
+    const Float px2 = p1 * q1 - p2 * q4 + p3 * q3;
+    const Float px3 = p2 * q1 + p1 * q4 - p3 * q2;
+    const Float px4 = p2 * q2 - p1 * q3 + p3 * q1;
     return Eigen::Vector3f(px2 * q1 - px1 * q2 - px3 * q4 + px4 * q3,
                            px3 * q1 - px1 * q3 + px2 * q4 - px4 * q2,
                            px3 * q2 - px2 * q3 - px1 * q4 + px4 * q1);
@@ -74,24 +74,24 @@ inline Eigen::Vector4f quat_conj(const Eigen::Vector4f &q) {
 }
 
 inline Eigen::Vector4f quat_exp(const Eigen::Vector3f &w) {
-    const double theta2 = w.squaredNorm();
-    const double theta = std::sqrt(theta2);
-    const double theta_half = 0.5 * theta;
+    const Float theta2 = w.squaredNorm();
+    const Float theta = std::sqrt(theta2);
+    const Float theta_half = 0.5 * theta;
 
-    double re, im;
+    Float re, im;
     if (theta > 1e-6) {
         re = std::cos(theta_half);
         im = std::sin(theta_half) / theta;
     } else {
         // we are close to zero, use taylor expansion to avoid problems
         // with zero divisors in sin(theta/2)/theta
-        const double theta4 = theta2 * theta2;
+        const Float theta4 = theta2 * theta2;
         re = 1.0 - (1.0 / 8.0) * theta2 + (1.0 / 384.0) * theta4;
         im = 0.5 - (1.0 / 48.0) * theta2 + (1.0 / 3840.0) * theta4;
 
         // for the linearized part we re-normalize to ensure unit length
         // here s should be roughly 1.0 anyways, so no problem with zero div
-        const double s = std::sqrt(re * re + im * im * theta2);
+        const Float s = std::sqrt(re * re + im * im * theta2);
         re /= s;
         im /= s;
     }
